+++ packages/openid-connect/CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
- - use `req.getHeader('origin')` to get a `Origin` header value
+ - use new method `req.getHeader('origin')` to get a `Origin` header value
+ - use new property `req.parsedParameters` instead of `queryString.parse(req.parameters)`
+
 
 ### Deprecated